Set up queries, unable to connect to DB ATM, commiting for progress tracking

+++ db/schema.sql
@@ -11,7 +11,7 @@ CREATE TABLE department (
 CREATE TABLE role (
     id INT AUTO_INCREMENT PRIMARY KEY,
     title VARCHAR(40) UNIQUE NOT NULL,
-    salary DECIMAL NOT NULL,
+    salary INT NOT NULL,
     department_id INT NOT NULL,
     INDEX dep_index (department_id),
     CONSTRAINT fk_department FOREIGN KEY (department_id) REFERENCES department(id) ON DELETE CASCADE
@@ -25,5 +25,6 @@ CREATE TABLE employee (
     manager_id INT NULL,
     PRIMARY KEY (id)
 );
+
